+++ bot/constants/messages.py
@@ -1,19 +1,31 @@
+START_BOT_DESCRIPTION_MESSAGE = (
+    "Привет! На связи команда {}\n"
+    "Мы создали этого бота для помощи всем действующим поставщикам "
+    "Wildberries.\n\n"
+    "Здесь вы можете:\n"
+    "- отследить позиции вашей карточки в выдаче.\n"
+    "- увидеть остатки товара по складам\n"
+    "- узнать остатки товара по размерам.\n\n"
+    "Для начала работы нажмите Start"
+)
 START_MESSAGE = (
-    "Привет! Чтобы воспользоваться ботом, нужно подписаться на наш"
-    "telegram канал"
+    "Привет! Чтобы воспользоваться ботом, нужно подписаться на наш "
+    "telegram канал {}"
+)
+SUCCESS_SUBSCRIBE_MESSAGE = (
+    "Спасибо за подписку! Вы можете воспользоваться меню бота:"
 )
 FALSE_SUBSCRIBE_MESSAGE = (
     "Вы не подписались на наш telegram канал"
 )
 HELLO_MESSAGE = (
-    "Добро пожаловать в No name Bot!\n"
+    "Добро пожаловать в {}!\n"
     "Узнайте на каких позициях находится ваш товар в поиске Wildberries."
 )
 POSITION_PARSER_MESSAGE = (
-    "Для определения позиции артикула отправьте сообщение в формате:\n"
-    "Артикул Поисковая фраза\n"
-    "Например:\n"
-    "36704403 футболка женская"
+    "Для определения позиции артикула отправьте сообщение в формате\n"
+    "'Артикул Поисковая фраза'\n\n"
+    "Например:\n36704403 футболка женская"
 )
 POSITION_PARSER_EXPECTATION_MESSAGE = (
     "Запрос с артикулом *{}* и поисковой фразой "
@@ -33,7 +45,6 @@ ACCEPTANCE_RATE_MESSAGE = "Выберите склад:\nНапример:\nАс
 ACCEPTANCE_RATE_ANSWER_MESSAGE = (
     "Коэффицианта приемки для склада {} составляет {}"
 )
-SUBSCRIPTIONS_MESSAGE = "Вы подписаны на позиции: {}"
-ERROR_MESSAGE = "Неизвестная команда, воспользуйтесь меню"
-
+SUBSCRIPTIONS_MESSAGE = "Ваши подписки: {}"
+UNKNOWN_COMMAND_MESSAGE = "Неизвестная команда, воспользуйтесь меню"
 TO_MAIN_MENU_MESSAGE = "Спасибо за подписку!"
